feat: add transformers module and update data processing logic
- Added new file `src/transformers.py` for data transformation utilities
- Modified `src/data_processing.py` to integrate new transformers
